Add files via upload
Initial commit where I made two video to video transitions and one that fades to black and back in. I'd like to eventual unify them so that they all have the pattern:

Initial video > intermediate > after video.

Right now the the intermediate is either a screenshot of the current one or it fade to black and comes back.